requirements.txt
# torch @ https://download.pytorch.org/whl/cpu/torch-2.1.1%2Bcpu-cp311-cp311-linux_x86_64.whl
# torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.16.1%2Bcpu-cp311-cp311-linux_x86_64.whl
torch @ https://download.pytorch.org/whl/cpu/torch-1.5.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.6.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
opencv-python-headless
scipy==1.10.1
numpy==1.24.1
Pillow==9.4.0
scikit-image==0.20.0
python-bidi
PyYAML
Shapely==2.0.1
pyclipper
ninja
